[REFACTOR]: 알림 전송 로직을 유스케이스에서 구현, Scheduler가 호출하도록 수정
- 테스트 REST API와 연동 목적

related-to: #11
